Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 void main(){
-    int a;
-    scanf("%d",&a);
-    printf("Area: %.2f",(float)(a*a*3.14));
+    float a;
+    scanf("%f",&a);
+    printf("Area: %.2f",(a*a*3.14));
 }
